normalize cache key for root in ExpandSchema
the cache is queried by normalized key.
but in ExpandSchema(), root was stored with un-normalized key.

this causes trouble on Windows which have drive letter as upper case in
absolute path.  when normalize this absolute, drive letter is converted
to lower case.  and it causes cache miss and ExpandSchema() failure.