Revert fix #146
A more thorough investigation on loading specs from a generated server
from paths containing special characters (such as blank spaces,
parenthesis) did not conclude that it was necessary to alter how cache
entries are matched.

Indeed, this change deeply altered how keys in the cache are managed
and was therefore at best, a risky bet.

* reverted commit e8e27ff
* added check to ensure that the root pseudo-document is preloaded in
  cache under all conditions
* added more unit tests to check the behavior of the normalizers against
  escaped characters
* closes #145

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>